% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gplates.R
\name{black_white}
\alias{black_white}
\title{Plot map as in plot_gplates with default black background and white landmasses}
\usage{
black_white(mya, base_url = "http://gws.gplates.org/")
}
\arguments{
\item{mya}{a numeric vector of length one designating age to reconstruct in millions of years ago}

\item{base_url}{The url to use; make sure it ends with a slash}
}
\value{
a plot of the plates and continents at the selected age
}
\description{
Plot map as in plot_gplates with default black background and white landmasses
}
