% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplates_topology.R
\name{gplates_plate_polygons}
\alias{gplates_plate_polygons}
\title{Reconstruct plate polygons}
\usage{
gplates_plate_polygons(age)
}
\arguments{
\item{age}{A numeric vector of length one indicating the geologic time to reconstruct.}
}
\value{
An S4 object of class SpatialPolygonsDataFrame
}
\description{
Reconstruct plate polygons
}
