% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplates_reconstruct.R
\name{gplates_reconstruct_coastlines}
\alias{gplates_reconstruct_coastlines}
\title{Reconstruct ancient coastlines}
\usage{
gplates_reconstruct_coastlines(age)
}
\arguments{
\item{age}{A numeric vector of length one indicating the geologic time to reconstruct.}
}
\value{
An S4 object of class SpatialPolygonsDataFrame
}
\description{
Reconstruct ancient coastlines
}
