% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplates_reconstruct.R
\name{gplates_reconstruct_static_polygons}
\alias{gplates_reconstruct_static_polygons}
\title{Reconstruct static polygons}
\usage{
gplates_reconstruct_static_polygons(age)
}
\arguments{
\item{age}{A numeric vector of length one indicating the geologic time to reconstruct.}
}
\value{
An S4 object of class SpatialPolygonsDataFrame
}
\description{
Reconstruct static polygons
}
