% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gplates.R
\name{land_sea}
\alias{land_sea}
\title{Plot map as in plot_gplates with user set land and sea colors}
\usage{
land_sea(mya, land = "gray", sea = "white", mapbackground = "white")
}
\arguments{
\item{mya}{a numeric vector of length one designating age to reconstruct in millions of years ago}

\item{land}{What color to paint the land}

\item{sea}{What color to paint the sea}

\item{mapbackground}{What color to paint the background (that which isn't the globe or plates)

Burn the sky, boil the sea -- you can't take my color palette from me!}
}
\value{
a plot of the plates and continents at the selected age
}
\description{
Plot map as in plot_gplates with user set land and sea colors
}
