% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gplates.R
\name{plot_gplates}
\alias{plot_gplates}
\title{Plot a map of selected age}
\usage{
plot_gplates(mya, polyoutline, polyfill, coastoutline, coastfill,
  plateoutline, platefill)
}
\arguments{
\item{mya}{a numeric vector of length one designating age to reconstruct in millions of years ago}

\item{polyoutline}{a character vector designating the color of the polygon outlines}

\item{polyfill}{a character vector designating the color of the polygons}

\item{coastoutline}{a character vector designating the color of the coast outlines}

\item{coastfill}{a character vector designating the color of the continents}

\item{plateoutline}{a character vector designating the color of the plate outlines}

\item{platefill}{a character vector designating the color of the plates}
}
\value{
a plot of the plates and continents at the selected age
}
\description{
Plot a map of selected age
}
