% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gplates.R
\name{plot_gplates_cartesian}
\alias{plot_gplates_cartesian}
\title{Plot a map of tectonic plates and boundries at any point in time of Earth's history keeping Cartesian coordinates}
\usage{
plot_gplates_cartesian(mya, polyoutline, polyfill, coastoutline, coastfill,
  mapoutline, mapbackground)
}
\arguments{
\item{mya}{a numeric vector of length one designating age to reconstruct in millions of years ago}

\item{polyoutline}{a character vector designating the color of the polygon outlines}

\item{polyfill}{a character vector designating the color of the polygons}

\item{coastoutline}{a character vector designating the color of the coast outlines}

\item{coastfill}{a character vector designating the color of the continents}

\item{mapoutline}{a character vector designating the color to ouline the map}

\item{mapbackground}{a character vector designating the color of the map background}
}
\value{
a plot of the plates and continents at the selected age
}
\description{
Plot a map of tectonic plates and boundries at any point in time of Earth's history keeping Cartesian coordinates
}
