% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplates_topology.R
\name{gplates_plate_boundaries}
\alias{gplates_plate_boundaries}
\title{Reconstruct plate boundaries}
\usage{
gplates_plate_boundaries(age, base_url = "http://gws.gplates.org/")
}
\arguments{
\item{age}{A numeric vector of length one indicating the geologic time to reconstruct.}

\item{base_url}{The url to use; make sure it ends with a slash}
}
\value{
An S4 object of class SpatialLinesDataFrame
}
\description{
Reconstruct plate boundaries
}
