% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplates_reconstruct.R
\name{gplates_reconstruct_point}
\alias{gplates_reconstruct_point}
\title{Reconstruct a point}
\usage{
gplates_reconstruct_point(lon, lat, age)
}
\arguments{
\item{lon}{A numeric vector of length one.}

\item{lat}{A numeric vector of length one.}

\item{age}{A numeric vector of length one indicating the geologic time to reconstruct.}
}
\value{
Coordinates
}
\description{
Reconstruct a point
}
